+++ GenNet_utils/LocallyDirected1D.py
